Testing SG Filter
More detailed testing and plot generation of the SG filter. SG filter works as expected, will produce a smoothed version of noisy given sufficiently large no. of data points. Derivative estimate is decent, but not sufficient for what I want. Do not think it is suitable for Ring Down Meas.